moved images module to tags module
