+++ tags/config.go
@@ -1,4 +1,4 @@
-package images
+package tags
 
 import (
 	"github.com/jasonrichardsmith/sentry/sentry"
@@ -9,5 +9,5 @@ type Config struct {
 }
 
 func (c *Config) LoadSentry() (sentry.Sentry, error) {
-	return ImagesSentry{}, nil
+	return TagsSentry{}, nil
 }
